+++ src/main/java/com/jandi/plan_backend/tripPlan/reservation/controller/ReservationController.java
@@ -2,8 +2,10 @@ package com.jandi.plan_backend.tripPlan.reservation.controller;
 
 import com.jandi.plan_backend.tripPlan.reservation.dto.ReservationReqDTO;
 import com.jandi.plan_backend.tripPlan.reservation.dto.ReservationRespDTO;
-import com.jandi.plan_backend.tripPlan.reservation.service.ReservationService;
+import com.jandi.plan_backend.tripPlan.reservation.service.ReservationQueryService;
+import com.jandi.plan_backend.tripPlan.reservation.service.ReservationUpdateService;
 import com.jandi.plan_backend.security.JwtTokenProvider;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -12,16 +14,13 @@ import java.util.Map;
 
 @Slf4j
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/trip/reservation")
 public class ReservationController {
 
-    private final ReservationService reservationService;
+    private final ReservationUpdateService reservationUpdateService;
+    private final ReservationQueryService reservationQueryService;
     private final JwtTokenProvider jwtTokenProvider;
-
-    public ReservationController(ReservationService reservationService, JwtTokenProvider jwtTokenProvider) {
-        this.reservationService = reservationService;
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
 
     @GetMapping("/{tripId}")
     public ResponseEntity<?> getReservation(
@@ -29,7 +28,7 @@ public class ReservationController {
             @RequestHeader("Authorization") String token
     ) {
         String userEmail = jwtTokenProvider.getEmail(token.replace("Bearer ", ""));
-        Map<String, Object> reservations = reservationService.getReservation(userEmail, tripId);
+        Map<String, Object> reservations = reservationQueryService.getReservation(userEmail, tripId);
         return ResponseEntity.ok(reservations);
     }
 
@@ -40,7 +39,7 @@ public class ReservationController {
             @RequestBody ReservationReqDTO reservedDTO
     ) {
         String userEmail = jwtTokenProvider.getEmail(token.replace("Bearer ", ""));
-        ReservationRespDTO savedReservation = reservationService.createReservation(userEmail, tripId, reservedDTO);
+        ReservationRespDTO savedReservation = reservationUpdateService.createReservation(userEmail, tripId, reservedDTO);
         return ResponseEntity.ok(savedReservation);
     }
 
@@ -51,7 +50,7 @@ public class ReservationController {
             @RequestBody ReservationReqDTO reservedDTO
     ) {
         String userEmail = jwtTokenProvider.getEmail(token.replace("Bearer ", ""));
-        ReservationRespDTO savedReservation = reservationService.updateReservation(userEmail, reservationId, reservedDTO);
+        ReservationRespDTO savedReservation = reservationUpdateService.updateReservation(userEmail, reservationId, reservedDTO);
         return ResponseEntity.ok(savedReservation);
     }
 
@@ -61,7 +60,7 @@ public class ReservationController {
             @RequestHeader("Authorization") String token
     ) {
         String userEmail = jwtTokenProvider.getEmail(token.replace("Bearer ", ""));
-        boolean isSucceed = reservationService.deleteReservation(userEmail, reservationId);
+        boolean isSucceed = reservationUpdateService.deleteReservation(userEmail, reservationId);
         return isSucceed ? ResponseEntity.ok("삭제되었습니다") : ResponseEntity.badRequest().build();
     }
 }
